scrapnow/requirements.txt
aiohttp==3.6.2
aiohttp-apispec==2.2.1
alembic==1.4.2
apispec==3.3.0
async-timeout==3.0.1
asyncpg==0.20.1
asyncpgsa==0.26.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.2
lxml==4.5.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.2
multidict==4.7.5
pendulum==2.1.2
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
pytzdata==2020.1
PyYAML==5.3.1
setproctitle==1.1.10
six==1.14.0
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.4
webargs==5.5.3
yarl==1.4.2
zipp==3.1.0